refactor: 당첨 결과 메시지를 열거형으로 관리
WinningResult.java
WinningResultMessageFormat.java
